+++ apps/web/src/app/[locale]/quiz/_components/CreateOrSolve/CreateOrSolveSection.tsx
@@ -1,8 +1,9 @@
 import { css } from '_panda/css';
 
-import { Background } from '@/app/[locale]/quiz/_common/BackGround';
-import { MobileLayout } from '@/app/[locale]/quiz/_common/MobileLayout';
-import SelectQuizType from '@/app/[locale]/quiz/_createOrSolve/SelectQuizType';
+import { Background } from '@/app/[locale]/quiz/_components/BackGround';
+import { MobileLayout } from '@/app/[locale]/quiz/_components/MobileLayout';
+
+import SelectQuizType from './SelectQuizType';
 
 const CreateOrSolveSection = () => {
   return (
