+++ apps/web/src/app/[locale]/quiz/_components/MobileLayout.tsx
