+++ apps/web/src/app/[locale]/quiz/create/page.tsx
@@ -1,4 +1,4 @@
-import CreateQuizSection from '@/app/[locale]/quiz/create/CreateQuizSection';
+import CreateQuizSection from '@/app/[locale]/quiz/create/_components/CreateQuizSection';
 
 async function CreateQuizPage() {
   return <CreateQuizSection />;
