+++ apps/web/src/app/[locale]/quiz/solve/_components/solving/QuizProgressBar.tsx
