+++ apps/web/src/app/[locale]/quiz/solve/_constants/solveQuiz.constants.ts
