+++ apps/web/src/app/[locale]/quiz/solve/page.tsx
@@ -1,5 +1,5 @@
-import { SolveQuizProvider } from '@/app/[locale]/quiz/solve/SolveQuizContext';
-import SolveQuizSection from '@/app/[locale]/quiz/solve/SolveQuizSection';
+import { SolveQuizProvider } from '@/app/[locale]/quiz/solve/_components/SolveQuizContext';
+import SolveQuizSection from '@/app/[locale]/quiz/solve/_components/SolveQuizSection';
 
 async function SolveQuizPage() {
   return (
